Fixed No need for ID when creating in DB Fixed an update contact_update_modal.html issue

+++ src/main/java/rme/project/Controller/ReservationController.java
@@ -57,7 +57,7 @@ public class ReservationController {
         reservation.setStartDate(LocalDate.parse(params.get("startDate")));
         reservation.setEndDate(LocalDate.parse(params.get("endDate")));
         reservation.setLocation(params.get("location"));
-        reservation.setKmFromOffice(Integer.parseInt(params.get("kmFromOffice")));
+        reservation.setKmFromOffice(Double.parseDouble(params.get("kmFromOffice")));
 
         reservationRepo.update(reservation);
